+++ CompulsoryCow.Meta/CompulsoryCow.Meta/Meta.Constructor.cs
@@ -5,7 +5,7 @@ namespace CompulsoryCow;
 
 public partial class Meta
 {
-    /// <summary>This method returns the default constructor for the given type &lt;TClass&gt;
+    /// <summary>This method returns the default constructor for the given <typeparamref name="TClass"/>.
     /// If there is no default constructor null is returned.
     /// </summary>
     /// <typeparam name="TClass"></typeparam>
@@ -15,4 +15,14 @@ public partial class Meta
         return typeof(TClass).GetConstructor(Type.EmptyTypes);
     }
 
+    /// <summary>This method returns teh default constructor for the given <paramref name="type"/>.
+    /// If there is no default constructor null is returned.
+    /// </summary>
+    /// <param name="type"></param>
+    /// <returns></returns>
+    public static ConstructorInfo? GetDefaultConstructor(Type type)
+    {
+        var res = type.GetConstructor(Type.EmptyTypes);
+        return res;
+    }
 }