+++ CompulsoryCow.Meta/CompulsoryCow.Meta/CHANGELOG.md
@@ -4,8 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+##[6.2.0]
+Add method `public static ConstructorInfo? GetDefaultConstructor(Type type)`.
+
+##[6.1.0]
+Add `Meta.GetClassesWithAttribute<TAttribute>(AppDomain appDomain) where TAttribute : Attribute` method.
+Add `Meta.GetDefaultConstructor<TClass>()` method.
+
 ## [6.0.0]
-Deprecated GetCallintMethod.
+Deprecated GetCallingMethod.
 Deprecated GetProperty.
 Instead use CallerMemberNameAttribute.
 Remove compile warnings.